requirement.txt
pandas==2.1.3
numpy==1.26.*
google-cloud-storage
pyspark
google-cloud-bigquery
fsspec
gcsfs